pyproject.toml
[project]
name = "world-cup-rag-chat-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "groq>=0.31.0",
    "ipykernel>=6.30.1",
    "ipython>=9.4.0",
    "langchain>=0.3.27",
    "langchain-huggingface>=0.3.1",
    "langgraph>=0.6.3",
    "langsmith>=0.4.13",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.1",
    "sentence-transformers>=5.1.0",
    "streamlit>=1.48.0",
]
